// src/store/modules/Facturas/enums/StoreEnums.ts
export enum Actions {
  GET_ALL_INVOICES = "getAllInvoices",
  SAVE_INVOICE = "saveInvoice",
  GET_INVOICE = "getInvoice",
  UPDATE_INVOICE = "updateInvoice",
  DELETE_INVOICE = "deleteInvoice",
}
  
export enum Mutations {
  SET_INVOICES = "setInvoices",
  SET_INVOICE = "setInvoice",
}